+++ crates/eww/src/widgets/widget_definitions.rs
@@ -819,14 +819,28 @@ fn build_gtk_label(bargs: &mut BuilderArgs) -> Result<gtk::Label> {
     def_widget!(bargs, _g, gtk_widget, {
         // @prop text - the text to display
         // @prop limit-width - maximum count of characters to display
+        // @prop truncate_left - whether to truncate on the left side
         // @prop show_truncated - show whether the text was truncated
-        prop(text: as_string, limit_width: as_i32 = i32::MAX, show_truncated: as_bool = true) {
-            let truncated = text.chars().count() > limit_width as usize;
-            let mut text = text.chars().take(limit_width as usize).collect::<String>();
-
-            if show_truncated && truncated {
-                text.push_str("...");
-            }
+        prop(text: as_string, limit_width: as_i32 = i32::MAX, truncate_left: as_bool = false, show_truncated: as_bool = true) {
+            let limit_width = limit_width as usize;
+            let char_count = text.chars().count();
+            let text = if char_count > limit_width {
+                let mut truncated: String = if truncate_left {
+                    text.chars().skip(char_count - limit_width).collect()
+                } else {
+                    text.chars().take(limit_width).collect()
+                };
+                if show_truncated {
+                    if truncate_left {
+                        truncated.insert_str(0, "...");
+                    } else {
+                        truncated.push_str("...");
+                    }
+                }
+                truncated
+            } else {
+                text
+            };
 
             let text = unescape::unescape(&text).context(format!("Failed to unescape label text {}", &text))?;
             let text = unindent(&text);